+++ shardingsphere-features/shardingsphere-sharding/shardingsphere-sharding-route/src/test/java/org/apache/shardingsphere/sharding/route/engine/validator/ddl/ShardingPrepareStatementValidatorTest.java
@@ -17,20 +17,21 @@
 
 package org.apache.shardingsphere.sharding.route.engine.validator.ddl;
 
+import com.google.common.collect.Sets;
 import org.apache.shardingsphere.infra.binder.statement.SQLStatementContext;
-import org.apache.shardingsphere.infra.binder.statement.ddl.AlterTableStatementContext;
+import org.apache.shardingsphere.infra.binder.statement.ddl.PrepareStatementContext;
 import org.apache.shardingsphere.infra.exception.ShardingSphereException;
 import org.apache.shardingsphere.infra.metadata.schema.ShardingSphereSchema;
 import org.apache.shardingsphere.infra.route.context.RouteContext;
 import org.apache.shardingsphere.infra.route.context.RouteMapper;
 import org.apache.shardingsphere.infra.route.context.RouteUnit;
-import org.apache.shardingsphere.sharding.route.engine.validator.ddl.impl.ShardingAlterTableStatementValidator;
+import org.apache.shardingsphere.sharding.route.engine.validator.ddl.impl.ShardingPrepareStatementValidator;
 import org.apache.shardingsphere.sharding.rule.ShardingRule;
-import org.apache.shardingsphere.sql.parser.sql.common.segment.ddl.constraint.ConstraintDefinitionSegment;
+import org.apache.shardingsphere.sql.parser.sql.common.segment.generic.table.JoinTableSegment;
 import org.apache.shardingsphere.sql.parser.sql.common.segment.generic.table.SimpleTableSegment;
-import org.apache.shardingsphere.sql.parser.sql.common.statement.ddl.AlterTableStatement;
 import org.apache.shardingsphere.sql.parser.sql.common.value.identifier.IdentifierValue;
-import org.apache.shardingsphere.sql.parser.sql.dialect.statement.postgresql.ddl.PostgreSQLAlterTableStatement;
+import org.apache.shardingsphere.sql.parser.sql.dialect.statement.postgresql.ddl.PostgreSQLPrepareStatement;
+import org.apache.shardingsphere.sql.parser.sql.dialect.statement.postgresql.dml.PostgreSQLSelectStatement;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.Mock;
@@ -57,65 +58,57 @@ public final class ShardingPrepareStatementValidatorTest {
     
     @Test
     public void assertPreValidatePrepareWithSameDatasourceSingleTablesForPostgreSQL() {
-        PostgreSQLAlterTableStatement sqlStatement = new PostgreSQLAlterTableStatement();
-        sqlStatement.setTable(new SimpleTableSegment(0, 0, new IdentifierValue("t_order")));
-        ConstraintDefinitionSegment definitionSegment = new ConstraintDefinitionSegment(0, 0);
-        definitionSegment.setReferencedTable(new SimpleTableSegment(0, 0, new IdentifierValue("t_order_item")));
-        sqlStatement.getAddConstraintDefinitions().add(definitionSegment);
-        SQLStatementContext<AlterTableStatement> sqlStatementContext = new AlterTableStatementContext(sqlStatement);
-        when(shardingRule.tableRuleExists(Arrays.asList("t_order", "t_order_item"))).thenReturn(false);
-        when(shardingRule.isSingleTablesInSameDataSource(Arrays.asList("t_order", "t_order_item"))).thenReturn(true);
-        new ShardingAlterTableStatementValidator().preValidate(shardingRule, sqlStatementContext, Collections.emptyList(), schema);
+        SQLStatementContext<PostgreSQLPrepareStatement> sqlStatementContext = new PrepareStatementContext(createPostgreSQLPrepareStatement());
+        Collection<String> tableNames = Sets.newHashSet("t_order", "t_order_item");
+        when(shardingRule.tableRuleExists(tableNames)).thenReturn(false);
+        when(shardingRule.isSingleTablesInSameDataSource(tableNames)).thenReturn(true);
+        new ShardingPrepareStatementValidator().preValidate(shardingRule, sqlStatementContext, Collections.emptyList(), schema);
     }
     
     @Test(expected = ShardingSphereException.class)
-    public void assertPreValidateAlterTableWithDifferentDatasourceSingleTablesForPostgreSQL() {
-        PostgreSQLAlterTableStatement sqlStatement = new PostgreSQLAlterTableStatement();
-        sqlStatement.setTable(new SimpleTableSegment(0, 0, new IdentifierValue("t_order")));
-        ConstraintDefinitionSegment definitionSegment = new ConstraintDefinitionSegment(0, 0);
-        definitionSegment.setReferencedTable(new SimpleTableSegment(0, 0, new IdentifierValue("t_order_item")));
-        sqlStatement.getAddConstraintDefinitions().add(definitionSegment);
-        SQLStatementContext<AlterTableStatement> sqlStatementContext = new AlterTableStatementContext(sqlStatement);
-        when(shardingRule.tableRuleExists(Arrays.asList("t_order", "t_order_item"))).thenReturn(false);
-        when(shardingRule.isSingleTablesInSameDataSource(Arrays.asList("t_order", "t_order_item"))).thenReturn(false);
-        new ShardingAlterTableStatementValidator().preValidate(shardingRule, sqlStatementContext, Collections.emptyList(), schema);
+    public void assertPreValidatePrepareWithDifferentDatasourceSingleTablesForPostgreSQL() {
+        SQLStatementContext<PostgreSQLPrepareStatement> sqlStatementContext = new PrepareStatementContext(createPostgreSQLPrepareStatement());
+        Collection<String> tableNames = Sets.newHashSet("t_order", "t_order_item");
+        when(shardingRule.tableRuleExists(tableNames)).thenReturn(false);
+        when(shardingRule.isSingleTablesInSameDataSource(tableNames)).thenReturn(false);
+        new ShardingPrepareStatementValidator().preValidate(shardingRule, sqlStatementContext, Collections.emptyList(), schema);
+    }
+    
+    private PostgreSQLPrepareStatement createPostgreSQLPrepareStatement() {
+        PostgreSQLSelectStatement select = new PostgreSQLSelectStatement();
+        JoinTableSegment joinTable = new JoinTableSegment();
+        joinTable.setLeft(new SimpleTableSegment(0, 0, new IdentifierValue("t_order")));
+        joinTable.setRight(new SimpleTableSegment(0, 0, new IdentifierValue("t_order_item")));
+        select.setFrom(joinTable);
+        PostgreSQLPrepareStatement sqlStatement = new PostgreSQLPrepareStatement();
+        sqlStatement.setSelect(select);
+        return sqlStatement;
     }
     
     @Test(expected = ShardingSphereException.class)
-    public void assertPostValidateAlterTableWithEmptyRouteResultForPostgreSQL() {
-        PostgreSQLAlterTableStatement sqlStatement = new PostgreSQLAlterTableStatement();
-        sqlStatement.setTable(new SimpleTableSegment(0, 0, new IdentifierValue("t_order")));
+    public void assertPostValidatePrepareWithEmptyRouteResultForPostgreSQL() {
+        PostgreSQLPrepareStatement sqlStatement = new PostgreSQLPrepareStatement();
         when(routeContext.getRouteUnits()).thenReturn(Collections.emptyList());
-        new ShardingAlterTableStatementValidator().postValidate(sqlStatement, routeContext);
+        new ShardingPrepareStatementValidator().postValidate(sqlStatement, routeContext);
     }
     
     @Test
-    public void assertPostValidateAlterTableWithSingleDataNodeForPostgreSQL() {
-        PostgreSQLAlterTableStatement sqlStatement = new PostgreSQLAlterTableStatement();
-        sqlStatement.setTable(new SimpleTableSegment(0, 0, new IdentifierValue("t_order")));
-        when(routeContext.getActualDataSourceNames()).thenReturn(Collections.singletonList("ds_0"));
+    public void assertPostValidatePrepareWithDifferentDataSourceForPostgreSQL() {
         Collection<RouteUnit> routeUnits = new LinkedList<>();
-        routeUnits.add(new RouteUnit(new RouteMapper("ds_0", "ds_0"), 
+        routeUnits.add(new RouteUnit(new RouteMapper("ds_0", "ds_0"),
                 Arrays.asList(new RouteMapper("t_order", "t_order_0"), new RouteMapper("t_order_item", "t_order_item_0"))));
         when(routeContext.getRouteUnits()).thenReturn(routeUnits);
-        new ShardingAlterTableStatementValidator().postValidate(sqlStatement, routeContext);
+        new ShardingPrepareStatementValidator().postValidate(new PostgreSQLPrepareStatement(), routeContext);
     }
     
     @Test(expected = ShardingSphereException.class)
-    public void assertPostValidateAlterTableWithMultiDataNodeForPostgreSQL() {
-        PostgreSQLAlterTableStatement sqlStatement = new PostgreSQLAlterTableStatement();
-        sqlStatement.setTable(new SimpleTableSegment(0, 0, new IdentifierValue("t_order")));
-        when(routeContext.getActualDataSourceNames()).thenReturn(Collections.singletonList("ds_0"));
+    public void assertPostValidatePrepareWithSameDataSourceForPostgreSQL() {
         Collection<RouteUnit> routeUnits = new LinkedList<>();
         routeUnits.add(new RouteUnit(new RouteMapper("ds_0", "ds_0"),
                 Arrays.asList(new RouteMapper("t_order", "t_order_0"), new RouteMapper("t_order_item", "t_order_item_0"))));
         routeUnits.add(new RouteUnit(new RouteMapper("ds_0", "ds_0"),
                 Arrays.asList(new RouteMapper("t_order", "t_order_0"), new RouteMapper("t_order_item", "t_order_item_1"))));
-        routeUnits.add(new RouteUnit(new RouteMapper("ds_0", "ds_0"),
-                Arrays.asList(new RouteMapper("t_order", "t_order_1"), new RouteMapper("t_order_item", "t_order_item_0"))));
-        routeUnits.add(new RouteUnit(new RouteMapper("ds_0", "ds_0"),
-                Arrays.asList(new RouteMapper("t_order", "t_order_1"), new RouteMapper("t_order_item", "t_order_item_1"))));
         when(routeContext.getRouteUnits()).thenReturn(routeUnits);
-        new ShardingAlterTableStatementValidator().postValidate(sqlStatement, routeContext);
+        new ShardingPrepareStatementValidator().postValidate(new PostgreSQLPrepareStatement(), routeContext);
     }
 }
